pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "govintel"
version = "0.1.0"
description = "Federal Procurement Intelligence Engine — RAG system analyzing federal contract data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Avishek Saha" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "chromadb>=0.5.0",
    "pinecone-client>=5.0.0",
    "sentence-transformers>=3.3.0",
    "rank-bm25>=0.2.2",
    "google-generativeai>=0.8.0",
    "ragas>=0.2.0",
    "langfuse>=2.0.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/govintel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
